// x-pack/platform/plugins/shared/cases/common/workflows/steps/add_alerts.d.ts
declare const InputSchema: z.ZodObject<{
    case_id: z.ZodString;
    alerts: z.ZodArray<z.ZodObject<{
        alertId: z.ZodString;
        index: z.ZodString;
        rule: z.ZodOptional<z.ZodObject<{
            id: z.ZodOptional<z.ZodString>;
            name: z.ZodOptional<z.ZodString>;
        }, z.core.$strip>>;
    }, z.core.$strip>>;
}, z.core.$strip>;
// ...
type AddAlertsStepInputSchema = typeof InputSchema;
type AddAlertsStepOutputSchema = typeof OutputSchema;
export type AddAlertsStepInput = z.infer<typeof InputSchema>;
export declare const addAlertsStepCommonDefinition: CommonStepDefinition<AddAlertsStepInputSchema, AddAlertsStepOutputSchema>;